Added chessboard view
